Extract Schema::Addition for running the traversal
